refactor push

+++ src/Main/HUD.java
@@ -4,17 +4,17 @@ import java.awt.*;
 
 public class HUD {
 
-    public static int HEALTH = 100;
+    public static float HEALTH = 100;
 
     private int greenValue = 255;
     private int score = 0;
     private int level = 1;
 
     public void tick() {
-        HEALTH = Game.clamp(HEALTH, 0, 100);
+        HEALTH = Game.clamp((int)HEALTH, 0, 100);
 
         greenValue = Game.clamp(greenValue, 0, 255);
-        greenValue = HEALTH * 2;
+        greenValue = (int)HEALTH * 2;
 
         score++;
 
@@ -24,7 +24,7 @@ public class HUD {
         g.setColor(Color.GRAY);
         g.fillRect(15, 15, 200, 32);
         g.setColor(new Color(75, greenValue, 0));
-        g.fillRect(15, 15, HEALTH * 2, 32);
+        g.fillRect(15, 15, (int)HEALTH * 2, 32);
         g.setColor(Color.WHITE);
         g.drawRect(15, 15, 200, 32);
